requirements.txt
PyYAML
jsonschema
fuzzywuzzy
python-Levenshtein

pdfplumber
pymupdf
pytesseract # also needs to be downloaded outside pip
Pillow
opencv-python
pdf2image
streamlit
